Fix unit tests after toolbox code move
