+++ test/unit/tools/test_toolbox.py
@@ -15,7 +15,7 @@ from galaxy.model import tool_shed_install
 from galaxy.model.tool_shed_install import mapping
 from galaxy.tools import ToolBox
 from galaxy.tools.cache import ToolCache
-from .test_toolbox_filters import mock_trans
+from ..tool_util.toolbox.test_toolbox_filters import mock_trans
 from ..tools_support import UsesApp, UsesTools
 from ..unittest_utils.sample_data import SIMPLE_MACRO, SIMPLE_TOOL_WITH_MACRO
